- Main: History for AddStoryDialog's Dashboard group ID combo box was being destroyed by explicitly setting the value in the QLineEdit.  Now the index is being set instead. - Main: Cached Reporter data has been moved out of the main config's "AddStoryDialog" section and into their own respective configuration files. This will facilitate easier data upgrading on start-up. - Main: Reporter parameter data have been partitioned into "Defaults" and "Stories" for independent file data storage.  This specific extrication and isolation of Reporter parameter data allows for upgrading that data (or downgrading it, as the case may be) on application start-up to ensure a sane operating environment. - Main: Updated the ReporterInfo metadata to include the Reporter's parameters version and requirements for convenience in saving/loading cached Reporter parameter settings (default and live) without having to create a temporary Reporter instance to query.
- Settings: Renamed the cache() method to init().
- Settings: Added a Boolean argument to the init() method to indicate that the
instance should be cleared of all settings.
- Settings: Added the get_filename() method for retrieving the actual filename
with format-specific extension applied.
- Settings: Added the index-based get_item() overload for retrieving data by
index within a Section instead of by name.

- IReporter: Changed the return value of the the Supports() method to be a
floating point value which should be between 0.0 and 1.0, representing "No, I
can't handle that" on the low end, and "That belongs exclusively to me" on the
high end.  Values in between are the plug-in's best guess as to how successful
it will likely be handling the Story.
- IReporter: Added a RequiresVersion() method to the interface so Reporter
requirements format can be versioned.
- IReporter: Added a RequiresFormat() method to the interface so Reporter
requirements format can be reportered.  This format is currently only the
original "Simple" format, but may be expanded in future versions to allow
Reporters to specify their requirements in other formats (XML, JSON, etc.).
- IReporter: Added a RequiresUpgrade() method to the interface so Reporter data
in previous formats can be upgraded to current formats.
- IReporter: Added a checkbox type to the Reporter requirements type.
- IReporter: Added a version parameter to the Requires() interface method for
managing parameter format versions.

- TeamCity9: Updated the Reporter's parameter requirements to include the
ability to control the querying of pending changes for a builder (which
highlighted the lack of a canonical way of upgrading plug-in parameters for
live Stories).
- TeamCity9: Added retrieval of the pending changes for an idle builder.  This
is now added to the "idle" headline with the prefix "Changes:" so it can be used
as a stylesheet trigger keyword.  As the change count increases, the Chyron will
update with the new value until the single-query limit is reached (100 in TC v9
API, for example), whereupon updates will stop for that builder until a build
is started.